fix(long-to-int) :sparkles: fixed issue toInt() in long value upper to limit

+++ app/src/main/java/com/rodrigo/deeplarva/domain/view/CameraValues.kt
@@ -9,7 +9,7 @@ data class CameraValues (
     var exposure: Int,
     val exposureMin: Int,
     val exposureMax: Int,
-    var shootSpeed: Int,
-    val shootSpeedMin: Int,
-    val shootSpeedMax: Int,
+    var shootSpeed: Long,
+    val shootSpeedMin: Long,
+    val shootSpeedMax: Long,
 )
